Create a minimal version of the library The printer's code general version has been moved to another file in the library.
